TRUNK-5388 Make ModuleFileParser MessageSourceService dependency explicit
new
* add constructor ModuleFileParser(MessageSourceService) to clearly
state dependencies, which enables easier testing by passing in a mock or
use of a different implementation of a MessageSourceService
* add method ModuleFileParser.parse(File)
* add method ModuleFileparser.parse(InputStream)
which provide the same functionality/use the same implementation
underneath

deprecate
* deprecate all other constructors (3 in total)
* deprecate parse() in favor of parse(File), there is no need to keep
the file as state. This way we can create one parser and reuse it by
just passing in a different file to parse

tests
* move all tests which do not need the Context to be there to the
ModuleFileParserUnitTest which is now BaseContextMockTest and uses mocks
of the MessageSourceService
* one test still needs to stay in the BaseContextSensitive test since
the Context.getLocale() is still called. Did not add Locale as
dependency since I think this specific method getMessage(String key,
Object[] args) should be implementated in the MessageSourceService
as is getMessage(String) which abstracts away the users current locale
for us
